Update button hover gradient
