+++ src/components/Footer.tsx
@@ -1,8 +1,10 @@
 const Footer = () => {
   const currentYear = new Date().getFullYear();
   return (
-    <footer className="flex justify-center items-end w-full py-6">
-      <p className="text-dark-blue ">&copy; {currentYear} Scout It Out</p>
+    <footer className="flex justify-center items-end w-full py-6 mt-4">
+      <p className="text-dark-blue dark:text-white">
+        &copy; {currentYear} – Scout It Out
+      </p>
     </footer>
   );
 };
